fix: policyname processing for kubernetes policies
Since all policy types are converted to Calico NetworkPolicy and GlobalNetworkPolicy
in the policy cache, we rely on the structure of the name to identify if a policy
is staged or is kubernetes or calico type.

kubernetes and staged kubernetes policy names can inlucde '.', whereas
calico policy types cannot have '.' in the name.

Hence when analyzing the name to identify the policy type, we need to consider
the "staged:" prefix in the beginning of the name for staged policies and then
look for 'knp', 'kanp', and 'kbanp' to identify kubernetes types.

in case of an unknown policy name structure we log error and continue with
the rest of the processing.

Tests are added accordingly to cover all the cases.