[TSLA-7610] Linseed logs reporting list and watch errors for *v1.Name… (#9086)
* [TSLA-7610] Linseed logs reporting list and watch errors for *v1.Namespace

This errors is reported by a shared informer that watches namespaces on
the managed cluster. When the managed cluster gets disconnected, it will
report this error and then try continously to restart the watch.

A disconnected cluster will be treated the same way as a deleted
cluster. Once the connection is re-established, the token controller
will receive an update and re-create the shared informer.

* [CODE REVIEW] Serialize managed cluster operation on the same channel