Fix queryserver to match policies without tier prefix for endpoints
temporary fix to allow policies without the tier prefix in the name (only applies to calico policies in default tier)
to be matched for endpoints. The reason we need to do this fix is that the selector from label inheritance returns
policies with the tier prefix and not their actual name as it is stored in the CRD. However, policy cache contains
the actual name stored and thus, keys.Contains(p.getKey()) would not be able to match policies created without
the default tier prefix.

This fix is adding the `default.` prefix if policy name does not include `.` which is correct as of now since
calico policies are not allowed to have `.` in the name except for when it is after tier name.