Update l7-collector to use same Envoy binary as envoy-proxy
Context: we do our own builds of the Envoy code so that we can apply any
patches that we need, and so that we have the ability to address CVEs if and
when they arise.  Our Envoy builds are managed by the
github.com/tigera/envoybinary repo, and the resulting images are published to
quay.io/tigera/envoybinary.