[CORE-11657] BPF: add packet rate QoS control (#10740)
* [CORE-11657] BPF: add packet rate QoS control

Add the BPF dataplane implementation of the packet rate QoS control.
Implement a token bucket filter (TBF) algorithm to limit ingress and/or
egress packet rate(s). Use the ifstate map to keep the persistent
parameters for the TBF. Add a UT and update QoS FVs for BPF coverage
of packet rate controls.

* read QoS from ifstate on bpfEndpointManager updates to persist QoS state correctly