[master] Replace the IPSetMember struct with an interface. (#9132)
Replace the IPSetMember struct with an interface.

The IPSetMember struct:

Was very large, with high occupancy (we store hundreds of thousands of them in larger clusters).
Contained lots of fields that were only used in a minority of cases.
Was error prone: which fields were allowed to be "zero" and in which combinations?
This PR replaces the struct with an interface along with several small implementation structs covering each use case. This reduces memory usage a lot and it makes it explicit which combinations of fields are required.

Move the protobuf encoding to the implementation structs; this is cleaner than a type switch in the event sequencer.

For the very common cases (simple IP-only or CIDR-only members), go the extra mile and make variants of the structs that store:

A single IPv4 CIDR
A single IPv6 CIDR
A single IPv4 addr
A single IPv6 addr
This means that, instead of a large struct embedding an ip.CIDR interface, which in turn points at a CIDR struct, we just have an interface that points at a CIDR (or an even smaller IP) struct.

Benchmarks show 10-20% reduction in calc graph memory usage in Enterprise:

* Fix inconsistencies in EGW IP set member type.

- On hitting (an unlikely) error when serialising dates, would
  return protobuf format with only 3 fields, not 5.
- Remove ToLower/ToUpper conversions.
- Treat empty string as zero time when parsing.