Fix flake in capture tests. (#9594)
Observed a nil pointer access for info.Size() in CI. Defend against
errors when reading the directory/converting to a file info object.

Rework the assertion function to do the read and assertions in a single
Eventually() block to avoid races.

Clean up some "defer ignores error" lint in the test file.