Force a fluent-bit linseed token refresh on 401 Unauthorized
Refresh the plugin token when a 401 Unauthorized is received.
This occurs when the external load balancer switches to a different
active cluster, causing the token to become invalid. A new token
for the non-cluster host service account will be requested from
the new active cluster. This changeset also bumps upstream
fluent-bit to v4.0.8 release.