Extract tigera-operator-signer public key from the CA bundle
Update the public key parsing logic to iterate through all certificates
in the CA bundle and return the public key for `tigera-operator-signer`,
rather than defaulting to the first certificate. This resolve an issue
where `tigera-operator-signer` is not the first certificate in the bundle.
This ensures that the Tigera authenticator can properly validate JWT
tokens issued by us.