[CI-1839] Improve /endpoints/aggregation response time (#9899)
Calls to /endpoints/aggregatio are executed by the UI when loading the
endpoints page. It retrieves flow logs that have denied traffic and all
endpoints via a call to the query server from the managed clusters.
The response from these two endpoints are merged together are returned
to the UI. When an user desired to show only endpoints with denied
traffic, the response from Linseed is fed into the request to the query
server. One proposed change is to execute these calls in parallel with
loading the page and execute them sequentially when we show only
endpoints with denied traffic.

Linseed request will return all flow logs that have denied traffic. If
the same endpoint has denied traffic every minute in the last 15
minutes, endpoints aggregation handler will extract a regex to match
endpoints containing the same endpoint enumerated 15 times. One proposed
change is use unique tokens when building the regex for denied
endpoints. Also, the regex used should use .*? instead of .* in order to
limit the string evaluation once it finds the namespace.

[CODE REVIEW] Track call duration and do not pass denied regex when
calls are made sequentially