[master] Maintain static ARP programming beyond endpoint start of day (#9864)
* Maintain static ARP programming beyond endpoint start of day

Before Calico v3.29 we refreshed the static ARP programming for a workload endpoint whenever the
endpoint interface state changed, or when the dataplane performed a full route refresh - which by
default means every 90 seconds (RouteRefreshInterval).  We accidentally lost that logic in
https://github.com/projectcalico/calico/pull/8418.  After that change, Calico only programs static
ARP when first configuring a workload endpoint, and does not refresh it after that point.

For most VMs (with Calico for OpenStack) and possibly all Kubernetes usage, that doesn't matter,
because the host kernel sends a dynamic ARP request when it needs to, and most VMs and pods respond
to that.  However it's possible for a VM to be configured such that it does not respond to such ARP
requests - e.g. with `arp_ignore` set to 2.  Connectivity to VMs like that worked before Calico
v3.29, but was broken from v3.29 onwards.

This PR reinstates the logic to refresh static ARP programming when the endpoint interface state
changes or when the dataplane performs a full resync.

* Add UT

* Rename "pendingARP" to "permanentARP"

* Improve commenting to cover dnsmasq's ARP programming