Simplify tunnel manager and use native go testing
This PR simplifies the tunnel manager for future changes by removing the interface type swithcing for generics and leveraging som chanutil tooling we have.

It also removes unecessary checks nad simplifies the state loop.

I've change the ginkgo tests to be go native tests as that's the direction we're moving in as well as used tunnel mocks for the tests to greatly simplify the tests.