Wait for status manager to finish closing before returning from voltron close
The tests are flaky because of a race condition where the next tests start before the status manager has completely shutdown from the previous test. A concurrent map iteration / write is occurring in the schema because the next tests add to the schema while the status manager reads from it

This commit ensures that the voltron server waits for the status manager to finish closing before returning, which ensures the next test doesn't start before the status manager closes.