Remove k8s token refresh test
The k8s token refresh test is testing k8s library functionality to ensure that if the token changes it is picked up. This is unecessary since we're using encapsulated third party code that hides the details of refreshing so that when Token is called it decides when it should recheck the file.

Using the NewCachedFileTokenSource is not option if a token path is given, so all we're doing is testing third party functionality works.