Remove tests that only test the fake k8s client
There are tests here that test that you can do CRUD operations with the fake k8s client and test nothing in voltron.

I imagine these were added to test the namespace vs cluster scoped ManagedCluster crds, but it doesn't even do that since 1) it's not installing any CRDs and 2) that stuff is mostly in the apiserver.