Fixes for 32-bit architectures (#11009) (#10250)
* Make IPNet test work on 32-bit architectures

Delegate the public NthIP method to a private one which accepts an
int64. Use that private method in the tests.



* Fix UID calculation on 32-bit architectures

The previous logic was truncating everything towards zero. Also fix some
trivial compile errors around integer sizes along the way.



* Fix compile errors around integer conversions on 32-bit



* Fix off-by-one error when calculating route table ID lengths



* Fix route table ID calculation on 32-bit architectures

When specifying a range that would exceed 32 (signed) bits when added
to the existing number of tables, the code silently wrapped around to
negative numbers instead of rejecting the range.



---------

Signed-off-by: Tom Wieczorek <twieczorek@mirantis.com>
Co-authored-by: Tom Wieczorek <twz123@users.noreply.github.com>