Separate out k8s logic from clusters, use channel for updates
This commit takes a step forward in simplifying / abstracting the k8s logic in the server / clusters code. It separates it out to the cluster / server code doesn't worry about setting up / managing k8s events directly, it just expects events to be sent over a provided channel.

One event, START_SYNC, on startup and if a resync is required. This happens in cases where we need to do a fresh list because we can't resume the watch. The receiver should block until it's retrieved all the managed cluster events and received the SYNC_STOP event.

Any clusters that weren't in this sync event must be removed.