Fix go routine leak caused by 10401
I removed the ServerStream wrapper in 10401 and replaced it with a go routine that's spawned per connection but only closed when the server context is cancelled.

The intent of the ServerStream wrapper was to ensure that we wait for all the connections to shutdown before we shutdown the server, however, a wrapper isn't needed for that. This required the Close was
called on the ServerStream, which isn't always guaranteed (so I think it's possible that we're leaking routines / not calling Done() on the WaitGroup).

The mux logic can signal when it shuts down with the CloseChan channel, so we do the following to guarantee we're not leaking routines / Done is always called when Add is called:
1. If there's an error before we successfully create a tunnel with the connection, just close the connection
2. If the tunnel is successfull created, immediately call s.wg.Go() with a select that waits for either tunnel.CloseChan to signal or the context to finish.